Разделил JavaScript код на модули
