+++ js/util.js
@@ -1,0 +1,15 @@
+const getRandomInRangeInt = (rangeFrom, rangeTo) => {
+  if (rangeFrom >= 0 && rangeFrom <= rangeTo) {
+    return Math.round(rangeFrom + (Math.random() * (rangeTo - rangeFrom)));
+  }
+  return null;
+};
+
+const getRandomInRangeFloat = (rangeFrom, rangeTo, accuracy) => {
+  if (rangeFrom >= 0 && rangeFrom <= rangeTo && accuracy >= 0) {
+    return (rangeFrom + (Math.random() * (rangeTo - rangeFrom))).toFixed(accuracy);
+  }
+  return null;
+};
+
+export {getRandomInRangeInt, getRandomInRangeFloat};
